fix(types): extract route helpers — Next.js 16 forbids non-route exports
Next.js 16 generates strict route-types in .next/types/* that fail with
TS2344 when route.ts / page.tsx files export anything besides recognised
route handlers. Two such helpers existed:

- app/api/ai/parse-cargo/route.ts exported MAX_EMAIL_BODY_CHARS,
  LLM_TIMEOUT_MS, PARSE_CARGO_CONCURRENCY, withRetry429, Retry429Options
- app/cargo/[id]/page.tsx exported renderSpecialRequirements

Extracted into dedicated modules:
- lib/parse-cargo-helpers.ts (parse-cargo constants + retry helper)
- lib/cargo-render.ts (specialRequirements normaliser)

Test imports updated to point at the new modules. Behaviour unchanged:
2167/2169 jest pass identical to main, npx tsc --noEmit now exits 0.

Pre-req for Wave γ — verify_commands include `npx tsc --noEmit` and
preflight would fail on these pre-existing errors.
