+++ app/cargo/[id]/page.tsx
@@ -12,34 +12,7 @@ import { cfValue } from '@/lib/types';
 import { AnalyticsTracker } from '@/lib/analytics-tracker';
 import { ClickableField } from '@/components/clickable-field';
 import { safeRender, getConf, ConfIcon } from '@/lib/ui-render';
-
-/**
- * βf2-02: Normalise specialRequirements before rendering.
- * The LLM parser sometimes returns an array of objects ({label, name, ...})
- * instead of the typed `string | null`. Coerce to readable text so the user
- * never sees "[object Object]" on the cargo page.
- *
- * Exported for unit testing (pure function, no React dependencies).
- */
-export function renderSpecialRequirements(
-  value: unknown,
-): string {
-  if (value == null) return '';
-  if (typeof value === 'string') return value;
-  if (Array.isArray(value)) {
-    if (value.length === 0) return '';
-    return value
-      .map((it) =>
-        typeof it === 'string'
-          ? it
-          : (it as Record<string, unknown>).label ??
-            (it as Record<string, unknown>).name ??
-            JSON.stringify(it),
-      )
-      .join(', ');
-  }
-  return safeRender(value as Parameters<typeof safeRender>[0]);
-}
+import { renderSpecialRequirements } from '@/lib/cargo-render';
 
 interface Props {
   params: Promise<{ id: string }>;
